requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.0.1
python-decouple==3.8
psycopg2-binary==2.9.7
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
